1. Добавил абстрактный класс
