+++ src/view/absract-view.js
@@ -1,0 +1,17 @@
+import { createElement } from '../render.js';
+
+export default class AbstractView {
+  #element = null;
+
+  get element() {
+    if (!this.#element) {
+      this.#element = createElement(this.template);
+    }
+
+    return this.#element;
+  }
+
+  removeElement() {
+    this.#element = null;
+  }
+}
