+++ src/view/content-list.js
@@ -1,25 +1,13 @@
-import { createElement } from '../render.js';
+import AbstractView from './absract-view';
+
 
 const createContentList = () => (
   '<ul class="trip-events__list"></ul>'
 );
 
-export default class ContentListView {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
+export default class ContentListView extends AbstractView{
 
   get template() {
     return createContentList();
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
