+++ src/view/sort-view.js
@@ -1,4 +1,4 @@
-import { createElement } from '../render.js';
+import AbstractView from './absract-view';
 
 const createSiteSortTemplate = () => (
   `<form class="trip-events__trip-sort  trip-sort" action="#" method="get">
@@ -30,22 +30,9 @@ const createSiteSortTemplate = () => (
 );
 
 
-export default class SiteSortView {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
+export default class SiteSortView extends AbstractView{
 
   get template() {
     return createSiteSortTemplate();
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
